+++ src/extension.ts
@@ -142,8 +142,13 @@ export function activate(context: vscode.ExtensionContext) {
     }
     //#endregion Region Status Bar
 
-    //#region Integration Events
+    //#region Special Copy
     context.subscriptions.push(vscode.commands.registerCommand('ats.copySelectionAsEventSubscriber', specialCopyFunct.copySelectionAsEventSubscriber));
+    context.subscriptions.push(vscode.commands.registerCommand('ats.copyAsRecordInsertStatement', specialCopyFunct.copyRecordInsertStatement));
+    context.subscriptions.push(vscode.commands.registerCommand('ats.copyAsRecordInsertStatementWithValidation', specialCopyFunct.copyRecordInsertStatementWithValidation));
+    context.subscriptions.push(vscode.commands.registerCommand('ats.copyAsRecordModifyStatement', specialCopyFunct.copyRecordModifyStatement));
+    context.subscriptions.push(vscode.commands.registerCommand('ats.copyAsRecordModifyStatementWithValidation', specialCopyFunct.copyRecordModifyStatementWithValidation));
+    context.subscriptions.push(vscode.commands.registerCommand('ats.copyAsRecordDeleteStatement', specialCopyFunct.copyRecordDeleteStatement));
 
     context.subscriptions.push(
         vscode.languages.registerCodeActionsProvider(
@@ -157,13 +162,13 @@ export function activate(context: vscode.ExtensionContext) {
 
     const generateSubscriberCommand = vscode.commands.registerCommand(
         'ats.copyAsEventSubscriber',
-        (alObject: ALObject, integrationEvent: string) => {
-            specialCopyFunct.copyAsEventSubscriber(alObject, integrationEvent);
+        (alObject: ALObject, sourceText: string) => {
+            specialCopyFunct.copyAsEventSubscriber(alObject, sourceText);
         }
     );
 
     context.subscriptions.push(generateSubscriberCommand);
-    //#endregion Integration Events
+    //#endregion Special Copy
 }
 
 export function deactivate() { }